fix(transform): infer fragment namespace by recursing into block children
`infer_namespace` only checked a fragment's DIRECT element children, so a
root/component/snippet fragment whose only `<svg>`/`<math>` element is nested
inside an `{#if}`/`{#each}`/`{#await}`/`{#key}` was mis-inferred as `html`,
emitting `$.from_html` instead of `$.from_svg` (plus the wrong inter-block
whitespace). Upstream's `infer_namespace` first runs `check_nodes_for_namespace`,
a recursive walk that descends through block containers and stops at the first
element — exactly the missing step.

Port `check_nodes_for_namespace` into the shared `3_transform/utils.rs` (the
faithful walk already lived privately in `snippet_block.rs`; that copy now
delegates to the shared one, removing the duplication) and call it from
`infer_namespace`.

Crucially, upstream only re-evaluates at namespace-RESET boundaries — where
`parent = path.at(-1) ?? node` is a Fragment/Root/Component/SvelteComponent/
SvelteFragment/SnippetBlock/SlotElement. rsvelte never populates `path`, so the
recursion is gated behind a new `is_reset_boundary` argument: the root component
fragment (`is_root_fragment`), component slot children, and snippet bodies pass
`true`; a fragment nested directly under an `{#if}`/element passes `false` and
keeps inheriting its element ancestor's namespace via the existing
direct-children check. Without this gate the walk would descend through a
sibling `{#if}` holding an `<svg>`/`<div>` and flip an element-scoped fragment's
namespace (caught two would-be regressions: flowbite AccordionItem, svelte.dev
Loading).

Verified green: byte-exact runtime/ssr/compiler_fixtures suites; corpus
verify.mjs reports no regressions.
